Add the default data directory to system/get/info command under key 'directory'

+++ src/TAO/API/commands/system/initialize.cpp
@@ -97,7 +97,7 @@ namespace TAO
             #endif
 
             /* Current unified time as reported by this node*/
-            jRet["timestamp"] =  (int)runtime::unifiedtimestamp();
+            jRet["timestamp"] = runtime::unifiedtimestamp();
 
             /* The hostname of this machine */
             jRet["hostname"]  = LLP::strHostname;
